requirements.txt
telethon
discord.py
python-dotenv
asyncio
flask